% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_benefit.R
\name{calculate_benefit}
\alias{calculate_benefit}
\title{Calculate ecosystem service benefit from a network}
\usage{
calculate_benefit(net, rival, alpha, beta, gamma, eta)
}
\arguments{
\item{net}{social-ecological network from which to calculate benefit (created using `network_create`)}

\item{rival}{whether the services are rival (TRUE) or non-rival (FALSE)}

\item{alpha}{the rate of production of the potential ecosystem service per unit area at supply nodes (value > 0)}

\item{beta}{the influence of connected supply nodes (i.e., the ecological-ecological links) on the rate of production of the potential ecosystem service at supply nodes (value > 0)}

\item{gamma}{marginal utility of the service at zero service used}

\item{eta}{rate of decline of the marginal utility with increasing use of the service (value > 0)}
}
\value{
A vector containing the ecosystem service benefit and the parameters used to generate the network
}
\description{
`calculate_benefit` calculates the ecosystem services benefits from a given
 network and a given set of rules.
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{social}
\keyword{spatial,}
\keyword{system,}
