% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ee_network.R
\name{create_ee_network}
\alias{create_ee_network}
\title{Create ecological-ecological networks}
\usage{
create_ee_network(ls_supply, ee_thresh, supply_area, e2e = TRUE, params = NULL)
}
\arguments{
\item{ls_supply}{Polygons containing ecosystem service supply areas}

\item{ee_thresh}{Distance threshold for the ecological-ecological links}

\item{supply_area}{Name of the column containing the supply area measure}

\item{e2e}{Logical. If `TRUE` edge-to-edge distances between patches are calculated. If `FALSE` centroid-to-centroid distances are calculated (the latter is much quicker)}

\item{params}{Vector containing the parameters used to generate the landscape if landscape is simulated (default = NULL)}
}
\value{
A list containing the network (and its attributes) and the parameters used to create the network
}
\description{
`create_ee_network` takes as input a spatial data layer (either real data or result of
 `ls_create`) and derives the underlying supply (ecological-ecological) network
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{spatial,}
\keyword{system,}
