% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_create.R
\name{ls_create}
\alias{ls_create}
\title{Create landscapes of ecosystem service supply (ecological) and demand (social) patches}
\usage{
ls_create(nrow, ncol, p_supply, p_demand, f_supply, f_demand, grad, inter)
}
\arguments{
\item{nrow}{Number of rows in the raster}

\item{ncol}{Number of columns in the raster}

\item{p_supply}{Proportion of the landscape which is ES supply}

\item{p_demand}{Proportion of the landscape which is ES demand}

\item{f_supply}{Fragmentation level of the ES supply (in the range [0, 1] with 1 being
the most fragmented)}

\item{f_demand}{Fragmentation level of the ES demand (in the range [0, 1] with 1 being
the most fragmented)}

\item{grad}{Steepness of spatial gradient for supply and demand (in the range [0, 1] with 1 being
the steepest gradient)}

\item{inter}{Interpersion between ES supply and demand along the spatial gradient (in
the range [0, 1] with 1 being most interspersed)}
}
\value{
A list containing a polygon containing the supply patches, a
 polygon containing the demand patches, and the parameters used to generate them
}
\description{
`ls_create` simulates ecosystem service supply and demand landscapes based
 on a number of parameters: proportion of supply/demand; fragmentation of
 supply/demand; steepness of supply gradient; interspersion of supply/demand
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{social}
\keyword{spatial,}
\keyword{system,}
