% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analyse.R
\name{network_analyse}
\alias{network_analyse}
\title{Create network summary statistics}
\usage{
network_analyse(net)
}
\arguments{
\item{net}{network object (output of `network_create`)}
}
\value{
Updated parameters with number and density of nodes
}
\description{
`network_analyse` takes as input a network object (output of `network_create`) and calculates the number and density of each type of nodes
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{social}
\keyword{spatial,}
\keyword{system,}
