% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_create.R
\name{network_create}
\alias{network_create}
\title{Create social-ecological networks from raster landscapes}
\usage{
network_create(ls, params = NULL, es_thresh, ee_thresh = NA,
  ss_thresh = NA)
}
\arguments{
\item{ls}{raster containing cells classified as supply, demand and neutral}

\item{params}{vector containing the parameters used to generate the landscape (if landscape is simulated)}

\item{es_thresh}{distance threshold for the ecological-social links}

\item{ee_thresh}{distance threshold for the ecological-ecological links (default value of `NULL` means no ecological-ecological links)}

\item{ss_thresh}{distance threshold for the social-social links (default value of `NULL` means no social-social links)}
}
\value{
A list containing the network (and its attributes) and the parameters used to create the network
}
\description{
`network_create` takes as input a raster (either real data or result of
 `ls_create` and derives the underlying social ecological network
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{social}
\keyword{spatial,}
\keyword{system,}
