% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_node.R
\name{raster_to_node}
\alias{raster_to_node}
\title{Create nodes from a raster}
\usage{
raster_to_node(ls, type_id)
}
\arguments{
\item{ls}{raster containing the supply-demand landscape}

\item{type_id}{Identifier for the cells to be converted to nodes}
}
\value{
An sf object containing nodes of the specified type. Area of each node is also calculated and added to the attributes
}
\description{
`raster_to_node` takes as input a raster layer and converts to nodes of one type
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{spatial,}
\keyword{system,}
