% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_benefit.R
\name{calculate_benefit}
\alias{calculate_benefit}
\title{Calculate ecosystem service benefit from a network}
\usage{
calculate_benefit(
  ee_network,
  es_network,
  rival,
  alpha,
  beta,
  gamma,
  params = NULL,
  lambda = 1,
  phi = 1
)
}
\arguments{
\item{ee_network}{Ecological-ecological network from which to calculate supply (created using `create_ee_network`)}

\item{es_network}{Social-ecological network from which to calculate benefit (created using `create_es_network`)}

\item{rival}{Whether the services are rival (TRUE) or non-rival (FALSE)}

\item{alpha}{The rate of production of the potential ecosystem service per unit area at supply nodes (value > 0)}

\item{beta}{The influence of connected supply nodes (i.e., the ecological-ecological links) on the rate of production of the potential ecosystem service at supply nodes (value > 0)}

\item{gamma}{Marginal utility of the service at zero service used}

\item{params}{Vector containing the parameters used to generate the landscape if landscape is simulated (default = NULL)}
}
\value{
A vector containing the ecosystem service benefit and the parameters used to generate the network
}
\description{
`calculate_benefit` Calculates the ecosystem services benefits from a given
 network and a given set of rules.
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{social}
\keyword{spatial,}
\keyword{system,}
