% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_es_network.R
\name{create_es_network}
\alias{create_es_network}
\title{Create social-ecological networks}
\usage{
create_es_network(ls_supply, ls_demand, es_thresh, supply_area_col = NULL,
  demand_area_col = NULL, params = NULL)
}
\arguments{
\item{ls_supply}{polygon containing ecosystem service supply areas}

\item{ls_demand}{polygon containing ecosystem service demand areas}

\item{es_thresh}{distance threshold for the social-ecological links}

\item{supply_area_col}{column which contains the area information for supply (NB this can be area, or modified area, such as by quality)

#'@param supply_area_col column which contains the area information for demand (NB this can be area, or modified area, such as by population)}

\item{params}{vector containing the parameters used to generate the landscape (if landscape is simulated, default = NULL)}
}
\value{
A list containing the network (and its attributes) and the parameters used to create the network
}
\description{
`create_es_network` takes as input a spatial data layer (either real data or result of
 `ls_create`) and derives the underlying supply (ecological-ecological) network
}
\keyword{ecological}
\keyword{ecosystem}
\keyword{landscape}
\keyword{model}
\keyword{neutral}
\keyword{services,}
\keyword{spatial,}
\keyword{system,}
